Add a new class DiscreteBarChart
Note the capital D. This is a new class which does not
render so many templates like discreteBarChart. Rather,
this class renders a templates which exploits Jinja
template inheritance. This template simply extends the template
content.html.

Future release can do:

    discreteBarChart = DiscreteBarChart

Such that, if discreteBarChart is removed the old class is still
available from the user's point of view.

Also, this class violates the DRY principle, it overrides methods
of NVD3Chart base class. This pattern will be introduced by all
other Capital letter new classes. This methods can be moved later to
a Mixin class.